# internal/api/spec/knot-openapi.yaml
openapi: 3.0.3
servers:
  - url: https://127.0.0.1
    description: Local server.
info:
  version: 1.0.0
  title: Knot API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  x-logo:
    url: ../images/logo-light.svg
  description: |
    # Introduction
    The REST API is used for the CLI, Web interface and 3rd party integrations.
...
paths:
...
  /api/users/{user_id}:
    get:
      tags:
        - Users
      summary: Get a User
      operationId: getUser
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user to get.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A user object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not-found'
      security: [BearerAuth: []]

    put:
      tags:
        - Users
      summary: Update User
      description: Update the user data.
      operationId: updateUser
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/userRequest'
      responses:
        '200':
          description: User updated successfully
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not-found'
      security: [BearerAuth: []]

    delete:
      summary: Delete User
      description: Delete a user, a user can't delete themselves.
      tags:
        - Users
      operationId: deleteUser
      parameters:
        - in: path
          name: user_id
          schema:
            type: string
            format: uuid
          required: true
          description: The ID of the user to delete.
      responses:
        '200':
          description: Successful operation.
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not-found'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]
  /api/users/{user_id}/quota:
    get:
      tags:
        - Users
      summary: Get User Quota
      operationId: getUserQuota
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user to get the quota for.
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: A user object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserQuota'
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not-found'
      security: [BearerAuth: []]
...
  /api/groups/{group_id}:
    put:
      summary: Update a Group
      description: Update a group.
      operationId: updateGroup
      tags:
        - Groups
      parameters:
        - name: group_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: The ID of the group to update.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GroupRequest'
      responses:
        '200':
          description: Successful operation
        '400':
          $ref: '#/components/responses/bad-request'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not-found'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]

    delete:
      summary: Delete Group
      description: Delete a specified group.
      tags:
        - Groups
      operationId: deleteGroup
      parameters:
        - in: path
          name: group_id
          schema:
            type: string
            format: uuid
          required: true
          description: The ID of the group to delete.
      responses:
        '200':
          description: Successful operation.
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not-found'
        '423':
          $ref: '#/components/responses/locked'
        '500':
          $ref: '#/components/responses/internal-server-error'
      security: [BearerAuth: []]

    get:
      summary: Get a Group
      description: Get a group.
      operationId: getGroup
      tags:
        - Groups
      parameters:
        - in: path
          name: group_id
          schema:
            type: string
            format: uuid
          required: true
          description: The ID of the group to get.
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupInfo'
        '401':
          $ref: '#/components/responses/unauthorized'
        '403':
          $ref: '#/components/responses/forbidden'
        '404':
          $ref: '#/components/responses/not-found'
      security: [BearerAuth: []]
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    GroupInfo:
      type: object
      properties:
        group_id:
          type: string
          format: uuid
          description: The ID of the group.
        name:
          type: string
          description: The name of the group.
        max_spaces:
          type: integer
          format: uint32
          description: The maximum number of spaces the group members can have.
        compute_units:
          type: integer
          format: uint32
          description: The number of compute units the group members get.
        storage_units:
          type: integer
          format: uint32
          description: The number of storage units the group members get.
        max_tunnels:
          type: integer
          format: uint32
          description: The maximum number of tunnels the group members can have.
...
    GroupRequest:
      type: object
      properties:
        name:
          type: string
          example: "my group"
          maxLength: 64
          description: The name of the group.
        max_spaces:
          type: integer
          format: uint32
          description: The maximum number of spaces the group members can have.
        compute_units:
          type: integer
          format: uint32
          description: The number of compute units the group members get.
        storage_units:
          type: integer
          format: uint32
          description: The number of storage units the group members get.
        max_tunnels:
          type: integer
          format: uint32
          description: The maximum number of tunnels the group members can have.
...
    userRequest:
      type: object
      properties:
        email:
          type: string
          maxLength: 255
          example: user1@example.com
          description: The email address of the user
        username:
          type: string
          maxLength: 64
          example: user1
          description: The username of the user
        password:
          type: string
          maxLength: 255
          example: password
          description: The password of the user
        service_password:
          type: string
          maxLength: 255
          example: password
          description: Option service password of the user, if empty a default one will be generated
        preferred_shell:
          type: string
          enum: [bash, zsh, fish, sh]
          description: The preferred shell of the user
        timezone:
          type: string
          description: The timezone of the user
        ssh_public_key:
          type: string
          description: The SSH key for the user
        github_username:
          type: string
          maxLength: 255
          description: The GitHub username of the user
        active:
          type: boolean
          description: Whether the user is active
        roles:
          type: array
          items:
            type: string
            format: uuid
            example: 00000000-0000-0000-0000-000000000000
          description: The IDs of the roles assigned to the user.
        groups:
          type: array
          items:
            type: string
            format: uuid
            example: 00000000-0000-0000-0000-000000000000
          description: The IDs of the groups assigned to the user.
        max_spaces:
          type: integer
          format: uint32
          description: The maximum number of spaces the user can have, 0 for unlimited.
        compute_units:
          type: integer
          format: uint32
          description: The number of compute units the user has.
        storage_units:
          type: integer
          format: uint32
          description: The number of storage units the user has.
        max_tunnels:
          type: integer
          format: uint32
          description: The maximum number of tunnels the user can have.
        totp_secret:
          type: string
          description: The TOTP secret for the user.
      required:
        - username
        - password
        - email
        - roles
    error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: The error message.
...
    UserQuota:
      type: object
      properties:
        max_spaces:
          type: integer
          description: The maximum number of spaces the user can have, 0 for unlimited.
        compute_units:
          type: integer
          format: uint32
          description: The number of compute units the user has.
        storage_units:
          type: integer
          format: uint32
          description: The number of storage units the user has.
        max_tunnels:
          type: integer
          format: uint32
          description: The maximum number of tunnels the user can have.
        number_spaces:
          type: integer
          description: The number of spaces the user has.
        number_spaces_deployed:
          type: integer
          description: The number of spaces the user has deployed currently.
        used_compute_units:
          type: integer
          format: uint32
          description: The number of compute units the user has used.
        used_storage_units:
          type: integer
          format: uint32
          description: The number of storage units the user has used.
        used_tunnels:
          type: integer
          format: uint32
          description: The number of tunnels the user has used.
    UserResponse:
      type: object
      properties:
        user_id:
          type: string
          format: uuid
          description: The ID of the user
        username:
          type: string
          description: The username of the user
        email:
          type: string
          description: The email of the user
        service_password:
          type: string
          example: password
          description: Option service password of the user, if empty a default one will be generated
        roles:
          type: array
          items:
            type: string
            format: uuid
          description: The roles of the user
        groups:
          type: array
          items:
            type: string
            format: uuid
          description: The groups of the user
        active:
          type: boolean
          description: Whether the user is active
        current:
          type: boolean
          description: Whether the user is the current user
        last_login_at:
          type: string
          format: date-time
          description: The last login time of the user
        created_at:
          type: string
          format: date-time
          description: The date and time the user was created
        updated_at:
          type: string
          format: date-time
          description: The date and time the user was last updated
        github_username:
          type: string
          description: The GitHub username of the user
        ssh_public_key:
          type: string
          description: The SSH key for the user
        preferred_shell:
          type: string
          enum: [bash, zsh, fish, sh]
          description: The preferred shell of the user
        timezone:
          type: string
          description: The timezone of the user
        max_spaces:
          type: integer
          description: The maximum number of spaces the user can have, 0 for unlimited
        compute_units:
          type: integer
          format: uint32
          description: The number of compute units the user has.
        storage_units:
          type: integer
          format: uint32
          description: The number of storage units the user has.
        max_tunnels:
          type: integer
          format: uint32
          description: The maximum number of tunnels the user can have.
        totp_secret:
          type: string
          description: The TOTP secret for the user.
...
  responses:
    bad-request:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    not-found:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    internal-server-error:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    locked:
      description: Locked
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
    unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/error'
...